change doubles and ints

+++ EAGRN-JMetal/src/main/java/eagrn/ComputingReferenceParetoFronts.java
@@ -67,8 +67,8 @@ public class ComputingReferenceParetoFronts {
 
         List<ExperimentProblem<DoubleSolution>> problemList = new ArrayList<>();
         for (int i = 0; i < networkFolders.length; i++) {
-            Map<String, Double[]> inferredNetworks = StaticUtils.readAllInferredNetworkFiles(files[i]);
-            CutOffCriteria cutOffCriteria = new PercLinksWithBestConfCriteria(0.4, geneNames[i]);
+            Map<String, Float[]> inferredNetworks = StaticUtils.readAllInferredNetworkFiles(files[i]);
+            CutOffCriteria cutOffCriteria = new PercLinksWithBestConfCriteria(0.4f, geneNames[i]);
             GRNProblem problem = new GRNProblem(inferredNetworks, geneNames[i], cutOffCriteria, strFitnessFormulas, strTimeSeriesFile);
             problem.setName(new File(networkFolders[i]).getName());
             problemList.add(new ExperimentProblem<>(problem));
